+++ express-server/.firebaserc
@@ -1,5 +1,5 @@
 {
   "projects": {
-    "default": "students-db-server"
+    "default": "students-server"
   }
 }
